Reorganise the code a bit
