+++ src/promptForEmail.ts
