+++ tests/aliasEmail.test.ts
@@ -1,6 +1,6 @@
 import { assertThrows, assert } from "https://deno.land/std/testing/asserts.ts";
 
-import aliasEmail from "../aliasEmail.ts";
+import aliasEmail from "../src/aliasEmail.ts";
 
 Deno.test("aliasEmail - should throw if invalid email provided", () => {
   assertThrows(() => aliasEmail("invalid"));
